fix an issue with hazelcast connection separation
