+++ docker/broker-registry.yml
@@ -28,7 +28,7 @@ services:
 
     activemq-broker:
         container_name: artemis-activemq-broker
-        image: docker.io/apache/activemq-artemis:2.39.0-alpine
+        image: docker.io/apache/activemq-artemis:2.41.0-alpine
         pull_policy: missing
         environment:
             ARTEMIS_USER: ${BROKER_USER:-guest}
